+++ CHANGELOG.md
@@ -14,6 +14,7 @@
      * `net\stubbles\lang\types\datespan\Month`
      * `net\stubbles\lang\types\datespan\Week`
      * `net\stubbles\lang\types\datespan\Year`
+   * removed `net\stubbles\lang\Clonable`
    * removed `net\stubbles\ioc\App::createPropertiesBindingModule()`, deprecated since 3.4.0
    * removed `net\stubbles\ioc\module\PropertiesBindingModule`, deprecated since 3.4.0
    * removed `net\stubbles\ioc\App::persistAnnotations()`, deprecated since 3.1.0